backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic-settings>=2.6.0
google-generativeai>=0.8.0
Pillow>=11.0.0
bcrypt>=4.2.0
PyJWT>=2.9.0
